Review the basic message.
Explain what the digest is exactly by referring to the original spec.

Using a [test
implementation](https://gitlab.com/FloweeTheHub/thehub/merge_requests/9)
I realized that the idea to send the intermediate state of the sha256 hash
(which is 64 bytes) doesn't make sense as we only save 64 bytes (2
intermediate sha256 hashes) doing so.

It is unfortunate that the initial replay-protection spec adds the
hashOutputs mid-state only so late to the hash, otherwise it would have
been very different.

Further;

Realize that since Schnorr we have multiple sizes of signatures. As
such we should specify the size in the message.

ps. before (with only ECDSA) we already had a variance of
signature-lenghts, but we could find out from the header of the
signature
how long it was. But IMOHO its better to just specify it.